backend/requirements.txt
# backend/requirements.txt

# Web Framework
fastapi
uvicorn[standard]
python-multipart

# AI & Embeddings
sentence-transformers
faiss-cpu
python-dotenv
pydantic-settings
nltk
numpy
pandas
pytesseract
joblib

# PDF Processing (from Round 1A)
PyMuPDF
scikit-learn
lightgbm

langchain==0.3.27  
langchain-openai==0.3.29  
langchain-google-genai==2.0.10  
langchain-community==0.3.27  
google-generativeai==0.8.5 
requests>=2.25.0
google-cloud-texttospeech==2.27.0
pydub>=0.25.0